% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrrstan_package.R
\docType{package}
\name{jrRstan-package}
\alias{jrRstan-package}
\alias{jrRstan}
\alias{rjrstan}
\title{R courses by Jumping Rivers}
\description{
Functions and data sets used in R courses by Jumping Rivers.
This package is used in the ``Introduction to Bayesian Inference using RStan''
course.
}
\keyword{package}
