% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nclRstan-package.R
\docType{package}
\name{nclRstan-package}
\alias{nclRstan-package}
\alias{nclRstan}
\alias{nclrstan}
\title{R courses by Jumping Rivers}
\description{
Functions and data sets used in R courses by Jumping Rivers.
This package is used in the ``Introduction to Bayesian Inference using RStan''
course.
}
\keyword{package}
