% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\docType{data}
\name{rats}
\alias{rats}
\title{Rat tumour data.}
\source{
The data are taken from Table 5.1 of Gelman \emph{et al.} (2013) but
were originally reported in Tarone (1982).
}
\usage{
data(rats)
}
\value{
A data frame with 71 rows and 2 variables. The data frame contains
the following columns:
\describe{
\item{y}{Number of rats in study with a tumour.}
\item{n}{Total number of rats in study.}}
}
\description{
Proportion of rats with tumours in 71 different studies.
}
\examples{
data(rats)
head(rats)
}
\references{
\insertRef{GCSDVR13}{jrRstan}

\insertRef{Tar82}{jrRstan}
}
\keyword{datasets}
