% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renal.R
\docType{data}
\name{renal}
\alias{renal}
\title{Survival times with right-censoring.}
\source{
The data set is taken from Henderson and Milner (1981).
}
\usage{
data(renal)
}
\value{
A data frame with 148 rows and 3 variables. The data frame contains
the following columns:
\describe{
\item{t}{The survival or censoring time, in months, of the patient.}
\item{status}{A factor with two levels, 0 and 1, where 0 indicates that the
patient's survival time was right-censored and 1 indicates that it was observed.}
\item{x}{The total number of HLA-B or DR antigen mismatches between the kindey
donor and recipient.}}
}
\description{
A data set containing the survival or right-censoring times of 148 renal
patients following a kidney transplant.
}
\examples{
data(renal)
head(renal)
}
\references{
\insertRef{HM91}{jrRstan}
}
\keyword{datasets}

