% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sideeffect.R
\docType{data}
\name{sideeffect}
\alias{sideeffect}
\title{Side effects data.}
\usage{
data(sideeffect)
}
\value{
A data frame with 7 rows and 3 variables. The data frame contains
the following columns:
\describe{
\item{dose}{Dose of drug.}
\item{n}{Total number of patients receiving drug.}
\item{effects}{Number of patients experiencing side effect.}}
}
\description{
Proportion of patients experiencing side effects after taking various dosages
of a drug for medical treatment.
}
\examples{
data(sideeffect)
head(sideeffect)
}
\keyword{datasets}

